Initial commit
"InternetGamesServer" is currently a simple WinSock server, which can accept many socket connections, each of which is handled in an individual thread.

Connecting from any of the Windows 7 Internet Games will reach only the "PLAY match" request phase, which is yet to be handled. A way to detect which of the 3 games is being played is needed as well, because of specific protocols.