Finding lobbies based on game, `Match`-es
`PlayerSocket` now searches for a lobby to join, depending on the game. If a free lobby doesn't exist, a new one is created. A lobby is essentially a `Match` with a state of `STATE_WAITINGFORPLAYERS`.

Searching for lobbies, based on a player's chosen skill level, is to be done.

To be done is also handling game-specific events within classes, derived from `Match`. Currently, the three `Match` arrays for the three games all use the `Match` base class directly.