`MatchManager`: Fix occasional crash when creating a match
Previously, the server had a good chance of crashing whenever a new match was being created.

The reason for that was the thread, which updates all matches, accessing the same vector a match is being pushed to at the same time.

A mutex was added to prevent this.

Additionally, `MatchManager` is now only updated once each second.