Checkers: Handling of "Resign" and "Draw" events
Game-specific matches can now access the data of "EventSend" messages and manipulate it for "EventReceive".

The `Match::QueuedEvent` struct allows for game-specific matches to indicate whether a resulting "EventReceive" should be sent to the sender as well.

Internet Checkers events, related to the "Resign" and "Draw" functions, are now handled.