Ending a game, when an opponent leaves
Receiving a "LEAVE" message from a client, no matter the state of `PlayerSocket`, will now disconnect it from the server.

When a `PlayerSocket` is to be disconnected, it's removed from its `Match`, if it is in one.

If the `Match` is in a `STATE_PLAYING` state and a player disconnects, the game is ended. This is because of missing game-specific logic on the server to allow for replacing with AI players.