Add Internet Games Client DLL project
The Internet Games Client DLL is to be injected into any of the 3 Internet Games, using an Injector application, which is currently yet to-be-done.

Currently, the DLL sets a breakpoint on the `GetAddrInfoW` WinSock function, and when that breakpoint is hit, it sets appropriate arguments, as well as pre-set remote host and port (which will be able to be specified from a dialog).