Client DLL: Set the required "DisableTLS" registry value to 1
The Client DLL now sets the "DisableTLS" registry value under "HKEY_CURRENT_USER\Software\Microsoft\zone.com\Zgmprxy" to 1, on process attach.

This value is required for connecting to the Internet Games Server, since it doesn't support TLS.

This addition to the Client DLL can mark it as completed (for now), since everything that was planned for it is now included.