Move `Match::ConstructGameInitXML()` to base `Match` class
The `Match::ConstructGameInitXML()` function is no longer virtual and is now defined in the `Match` class itself.

A virtual `Match::AppendToGameInitXML()` function was added to allow adding additional data. It's called from `ConstructGameInitXML()`.

This change was done to reduce code duplication, as the "GameInit" messages are pretty much the same between games.