Remove try/catch statements regarding `std::stoi` in `ProcessEvent()` functions
`BackgammonMatch` and `CheckersMatch` used to include try/catch statements to prevent an exception thrown from invalid `std::stoi`. However, if such exception were to happen, it would be the cause of modified/corrupt data from a client. The exception would be caught in `Socket::SocketHandler` and the client sending invalid data will be disconnected, ending the match.