Cleaner console output, print Socket events to log files ("--log")
The new "--log" (or "-l") argument allows for saving Socket event (send/receive) logs to files, either in the default "InternetGamesServer_logs" directory, or in a custom one (specified after the argument).

The console output is now cleaner and the beginning of each message contains specification on which module/class it's related to (MAIN, SOCKET, MATCH, MATCH MANAGER).

Additionally, `Socket` is now a class, instead of a namespace.