`SpadesMatch::UpdateTeamPoints()`: Award overtrick bonus points to a team, even if their contract wasn't met
Previously, awarding overtrick bonus points only occurred if a contract was met, which caused issues because contract fulfillment is based on non-Nil tricks only.

Overtrick bonus points are awarded based on total tricks taken by the team, so this shouldn't depend on the contract met check.

Additionally, add a debug macro definition to allow printing data about points/bags calculation for a team when a hand ends, to the console.