Client DLL: Custom host and port support for Windows XP Internet Games
If the `XP_GAMES` preprocessor macro is defined, the client DLL will be compiled for Windows XP Internet Games instead of Windows 7 ones.

This is utilized in a "InternetGamesClientDLL_XP.vcxproj.vcxproj" project, which can be found under the "InternetGamesClientDLL" directory (same as original client DLL project). The only difference with the new project file is that it defines the `XP_GAMES` preprocessor macro. The new project is built into a new "InternetGamesClientDLL_XP.dll" file, if the target architecture is x86.

Additionally, x86 builds of the DLL Injector now support the "-x" ("--xp") parameter, which indicates a Windows XP Internet Game is to be targeted. This works for all games, because they are all hosted by "zClientm.exe".