`WinXP::PlayerSocket`: Send additional "ServiceInfo" messages after "Hello", track "sequenceID" for sent generic messages
Additional required "ServiceInfo" messages are sent after the "Hello" message in `WinXP::PlayerSocket::SendProxyHelloMessages()`, as the full "serviceInfo" (which is the first of the two messages) is what triggers a state change in the client.

Additionally, the "sequenceID" of the `PlayerSocket` is incremented by 1 on each generic message sent, as to comply with the protocol.