Server: `WinXP::PlayerSocket`: Ensure game messages are only processed/sent in `STATE_PLAYING`
Fixes an issue, where sometimes one of the clients will prompt the match to broadcast its check-in message sooner than the other has sent over its game start messages, causing the game to never be initialized on the late client.

An event was added to dispatch whenever the state of the `WinXP::PlayerSocket` is `STATE_PLAYING`, as to ensure game messages are accepted. Assertions were additionally added.