Server: Always `ReleaseMutex()` if an exception is thrown, before re-throwing it
For code that already did this, the additional exception thrown for `ReleaseMutex()` failure was now removed in favour of always re-throwing the original exception. `catch (...)` is now also used as to ensure the mutex is released when anything is thrown.

Additionally, the `WinXP::PlayerSocket::AwaitIncomingGameMessage()` functions now always check if the game message type matches before checking for length match for more on-point exception messages.