Compile XP client DLL project with v141_xp toolset, add new XP DLLInjector project
The existing client DLL project for Windows XP is now compiled with the v141_xp toolset, allowing for it to be ran on Windows XP.

A new DLLInjector project for Windows XP (using the same Main.cpp file) has been added, which also compiles with the v141_xp toolset and a `WIN_XP` preprocessor define.

TODO: Support v141_xp toolset in the GitHub Actions workflow, if possible.