Server: WinXP: Provide "Powered by ZoneInternetGamesServer" ad banner
Whenever a Windows XP Internet Game, which has the custom DLL injected, requests an ad banner via HTTP, the server will return a simple HTML page, which contains a banner image that will be returned by the server as well (due to all hosts/ports being overriden with the specified ones from the DLL).

An argument was added to disable this behaviour and ignore such requests.

The custom "ad banner" is stored as a resource inside of the target server executable.