Server: WinXP: Proper chat message validation with support for other languages
The previous chat validation, which featured checking against each full chat message string individually, did not support other languages and thus marked such messages as invalid.

The new validation checks if the message string starts with a valid ID in the format "/{id}" and whether it fits the common/game-specific message range.
XP games initially check for a wide character at the end of the message string, which is equal to the message ID. So after the function returns a vaild ID, a message with a 4-byte array is broadcasted, one of the bytes in it being the ID.