Server: WinXP: Remove players from match without disconnecting them from the server
Whenever a player leaves a match, other players are now removed from the match but not disconnected from the server. They are left in a "disconnected proxy" state.

This change not only solves a race condition when it comes to disconnecting more than 2 clients from a match at the same time, but also prevents the "Error communicating with server" screen and its 5-second hang occuring. Instead, a "Your opponent has left the game" message is shown.