Docs: Update server and Windows XP Backgammon and Checkers preview images
The old server preview image shows an outdated version of the server which prints logs to the output stream.

The old Windows XP Backgammon and Checkers preview images were screenshots from Windows 7. For consistency with the other Windows XP game screenshots, the new ones are made on XP as well.