Server: `Win7::BackgammonMatch`: Improvements to message validation, detect game over
Various improvements to message validation were done, for example tracking whether the "Crawford Rule" is enabled for the current game and forbidding doubles.

The position of all stones on the board is now also tracked, as its required that we know if a losing opponent's piece is at the other player's home table or at the bar, since this is a requirement for giving the winner "Backgammon" (3) points from a game. We have to track points from games to be able to detect game over (after a player scores 5 points in the match).