WinME VBScript patcher: Avoid byte-writing workaround which invalidates certain bytes on WinME
Previously, some port numbers could not be properly written to CMNCLIM.DLL, due to certain bytes being represented as characters, unsupported by the "iso-8859-1" charset on Windows ME, and being automatically converted to binary as '?' (3F).

The script now uses `ADOBD.Stream.Write` to write the port as binary data. Obtaining a SAFEARRAY that is required for that function is done through a workaround involving decoding hex to bytes using MSXML, as VBScript cannot natively create a SAFEARRAY that works with `ADOBD.Stream.Write`.

Additionally, the hostname is no longer parsed to hex, as it's directly written as-is to each game executable via a text `ADOBD.Stream` through `ADOBD.Stream.WriteText`.

Some slight README tweaks are included with this commit.