Server: Fix issues and improve usage of mutexes to better prevent race-conditions
This is a lengthy commit, which does multiple changes to the overall usage of mutexes in the server.

Firstly, both `Win7` and `WinXP`'s `Match` classes now have only 1 general `m_mutex`. It is used for ensuring only one process in the match (regardless of the function) is done at a time. This should prevent many race conditions.

`MatchManager`'s usage of mutexes was extended to `DestroyPlayer`. `MatchManager` now also waits for an individual `Match`'s mutex, if it's about to be erased.

Every mutex/event on the server is affected by socket recv/send times, so now they all will have to wait the usual recv/send timeout time of 60 seconds + 10 more just in case.

Everywhere in the server, mutexes are all checked for `WAIT_OBJECT_0`, `WAIT_TIMEOUT`, `WAIT_ABANDONED` and `default` (any other wait state). This ensures consistency and will likely help with debugging.

Additionally, `WinXP::PlayerSocket::m_inLobby` is renamed to `m_proxyConnected`, as that name better outlines what that boolean is about.