Server: Add config option to limit number of clients from given IP
Additionally, `Socket::Socket()` is no longer provided a reference to a log stream, but rather it's created after construction in `Socket::Initialize()`. This should lead to a faster creation of a `Socket` object upon starting a thread with `Socket::SocketHandler()`, and thus, less time will be needed for the socket to be added to the list, where it can be discovered.